Cast away char subscripts to avoid warnings from clang.

+++ src/StringList.cxx
@@ -57,11 +57,11 @@ static char **ArrayFromStringList(char *stringList, int *len, bool onlyLineEnds 
 	for (int i=0; i<256; i++) {
 		wordSeparator[i] = false;
 	}
-	wordSeparator['\r'] = true;
-	wordSeparator['\n'] = true;
+	wordSeparator[static_cast<unsigned int>('\r')] = true;
+	wordSeparator[static_cast<unsigned int>('\n')] = true;
 	if (!onlyLineEnds) {
-		wordSeparator[' '] = true;
-		wordSeparator['\t'] = true;
+		wordSeparator[static_cast<unsigned int>(' ')] = true;
+		wordSeparator[static_cast<unsigned int>('\t')] = true;
 	}
 	for (int j = 0; stringList[j]; j++) {
 		int curr = static_cast<unsigned char>(stringList[j]);
@@ -178,7 +178,7 @@ bool StringList::InList(const char *s) {
 			j++;
 		}
 	}
-	j = starts['^'];
+	j = starts[static_cast<unsigned int>('^')];
 	if (j >= 0) {
 		while (words[j][0] == '^') {
 			const char *a = words[j] + 1;
